Ensure home activity always is on home stack.
It is possible to start a home activity on a non-home stack. For
example, an activity that handles the home intent may be started with
via a different intent, landing it on the fullscreen stack. This
becomes problematic later on due to our assumptions about the home
stack state and handling of tasks within it.

To ensure proper behavior, this changelist moves the activity
handling the home intent to the home stack if it is not already
present there.

Fixes: 36606833
Test: Move home activity to fullscreen stack, launch another
      activity, press home button.
Test: go/wm-smoke

Change-Id: If88550a24dca3eabc0c60940937e6a1e6f133cae