Don't move stack to bottom when creating task at bottom
When creating a task, only move the parents of the task if we are creating
the task at the top of the stack. i.e. creating the task at the top of its
stack will move the stack forward and creating the task at the bottom of
the stack shouldn't move the stack backwards. Creating a task at the bottom
is used for things like restoring a task from recents to use for other things
and for background user, so shouldn't affect visibility of its parent stack.

Change-Id: Ifcaf7c5a95294631dff675ac5c7d87f059b1792a
Fixes: 62352722
Test: go/wm-smoke
(cherry picked from commit 39083ba29df566dd06bb1711d5e06f9fb305afd2)