AOD: Fix more flickering
Fixes an issue where waking up with fingerprint would not work correctly
if the keyguard was not locked.

Also fixes flickering when triggering the Assistant on AOD.

To achieve the above, also ensures that whenever the scrim is visible, the status bar window is expanded.

Finally, log the alpha and tint of the scrims to systrace.

Fixes: 63983663
Fixes: 63065774
Bug: 63531607
Test: Disable power key locks immediately, trigger AOD, unlock with fingerprint. Verify no flicker; Trigger Assistant from AOD in various scenarios. Verify no flicker.

Change-Id: I0db2938afdde23c1995f6ac905d4a6abdf25cdec