AOD: Fix janky pulse out animation
Fixes an issue where during the pulse out animation, the fully
black frame did not hit the framebuffer before setting the power
state, causing visible jank in the ON -> DOZE transition.

To work around this, we now delay finishing the pulse and thus
setting the power state until the frame had time flush.

Change-Id: Ic1a04920fed516170d6803327cc9d46ed07e8b49
Fixes: 64277193
Test: Recieve notification on AOD, swipe away, verify no flicker (before fix reproduced ca 1/10)