Fix light bar flickers
There was a time in which no window was mTopFullscreenOpaqueWS
when switching between two apps because the previous one was
exiting already before the next one was shown. Fix this by only
considering a window to be exiting after the animation actually
has started. In that case the surface of the incoming window will
be visible already.

Test: go/wm-smoke
Test: Open Cast settings, go home with light wallpaper set
Change-Id: I89a85d5fbfe1618940b398cd26a8d5fe53be12fb
Fixes: 64291272