Fixes recents entrance animation when launching before onStop happens
If RecentsActivity was launched before its onStop runs then it would
have skipped the entrance animation because its launch flags were reset
in onStop. To trigger this, going home and immediately pressing recents
will bring it back up. To fix this, skip reset launch flags (because it
is set before onStop runs) when recents is requested to launch again.

Change-Id: I768f48c33fe77ab2e8f6957611930bf7605bc674
Fixes: 65205705
Test: manual - run recents with apps, press home and immediately recents